requirements.txt
openpyxl
pymupdf
pytesseract
pillow
python-docx
